refactor(bedrock): split provider tests for GH727
